Add tctl api-resources command
